Removed datefinder
it was rewriting the date format, despite me setting the 'base_date' attribute. it would rewrite MM/DD/YYYY into YYYY-MM-DD, counter intuitive to me wanting to later strip the found date from a string. Instead I've rewritten my 'find date' logic to just iterate over all chars in a string, and exit on the last char in the date sring, ':', but only when it's found for a 3rd time, it appears twice between hours:minutes and minutes:seconds