newfound bug while manually testing
manually testing, ew, I know. Just hard to replicate the target env and test against it. Well maybe not all that hard, but alot of effort I'm not going to put forth, yet. Maybe in the future should this come together and work as I envision.
When a player died in the server, an on_disconnect message was sent? Thank god for systemd on the server restarting bytebot. As a temp work around I'm